+++ test/Resources/abstract/ChildClass.js
@@ -1,0 +1,3 @@
+import BaseClass from './BaseClass'
+
+export default class ChildClass extends BaseClass {}